requirements.txt
streamlit
tensorflow
mysql-connector-python
pillow
matplotlib
pandas
